% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_type.R
\name{pb_var_to_type}
\alias{pb_var_to_type}
\title{Convert a pb_variable to an ellmer type}
\usage{
pb_var_to_type(var)
}
\arguments{
\item{var}{A \code{pb_variable} object}
}
\value{
An ellmer type object
}
\description{
Convert a pb_variable to an ellmer type
}
\keyword{internal}
